Fix #64 regarding KDocs
### What's done:
* Added tests on FunctionASTNodeUtils
